Renamed the component *-hoc files to *-container files.
